Rename legacy binaries
