+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "ptx-linker"
-version = "0.8.3"
+version = "0.9.0-alpha"
 authors = ["Denys Zariaiev <denys.zariaiev@gmail.com>"]
 license = "MIT"
 
@@ -9,34 +9,54 @@ description = "NVPTX modules linker"
 repository = "https://github.com/denzp/rust-ptx-linker"
 categories = ["development-tools", "development-tools::build-utils", "external-ffi-bindings"]
 keywords = ["llvm", "cuda", "nvptx", "linker"]
-autotests = true
+
+[badges]
+appveyor = { repository = "denzp/rust-ptx-linker", branch = "master", service = "github" }
+travis-ci = { repository = "denzp/rust-ptx-linker", branch = "master" }
+maintenance = { status = "experimental" }
+
+[features]
+default = ["llvm-proxy", "cli-legacy"]
+cli-legacy = []
+cli-rustc = []
+llvm-proxy = ["rustc-llvm-proxy"]
+llvm-external = []
 
 [[bin]]
-name = "ptx-linker"
+name = "legacy-ptx-linker"
 doc = false
-path = "src/application/main.rs"
+path = "src/bin/legacy-ptx-linker/main.rs"
+required-features = ["cli-legacy"]
 
 [[bin]]
-name = "rustc-dylib-wrapper"
+name = "legacy-rustc-dylib-wrapper"
 doc = false
-path = "src/rustc-dylib-wrapper/main.rs"
+path = "src/bin/legacy-rustc-dylib-wrapper/main.rs"
+required-features = ["cli-legacy"]
 
-[lib]
-name = "ptx_linker"
+[[bin]]
+name = "rust-ptx-linker"
+doc = false
+path = "src/bin/rust-ptx-linker/main.rs"
+required-features = ["cli-rustc"]
 
 [dependencies]
-log = "0.4"
-fern = "0.5"
-colored = "1.5"
-error-chain = "0.12"
 ar = "0.6"
-clap = { version = "2.31", features = ["yaml", "suggestions", "color"] }
-rustc-llvm-proxy = "0.1"
 either = "1.5"
+env_logger = "0.6"
+log = "0.4"
+clap = "2.31"
+
+# TODO: get rid of me!
+error-chain = "0.12"
 
 [dependencies.llvm-sys]
 version = "60"
 features = ["no-llvm-linking", "disable-alltargets-init"]
+
+[dependencies.rustc-llvm-proxy]
+version = "0.1"
+optional = true
 
 [dev-dependencies]
 failure = "0.1"
@@ -45,8 +65,3 @@ crate-compile-test = "0.2"
 [[test]]
 name = "linker"
 harness = false
-
-[badges]
-appveyor = { repository = "denzp/rust-ptx-linker", branch = "master", service = "github" }
-travis-ci = { repository = "denzp/rust-ptx-linker", branch = "master" }
-maintenance = { status = "experimental" }
