--- cli.yml
@@ -1,68 +1,0 @@
-name: "ptx-linker"
-about: Performs linking of Rust NVPTX crates
-
-args:
-  - input:
-      help: Path to input rlib or bitcode
-      multiple: true
-
-  - output:
-      short: o
-      help: Sets path for assembly output
-      takes_value: true
-      value_name: FILE.ptx
-
-  - optimise:
-      short: O
-      help: Sets optimisation level
-      takes_value: true
-      value_name: level
-
-  - __ignored_lib_start:
-      long: whole-archive
-      hidden: true
-      multiple: true
-
-  - __ignored_lib_end:
-      long: no-whole-archive
-      hidden: true
-      multiple: true
-
-  - __ignored_start_group:
-      long: start-group
-      hidden: true
-      multiple: true
-
-  - __ignored_end_group:
-      long: end-group
-      hidden: true
-      multiple: true
-
-  - __ignored_compiler_prefix:
-      short: B
-      takes_value: true
-      hidden: true
-      multiple: true
-      number_of_values: 1
-
-  - __ignored_shared_hack:
-      short: s
-      hidden: true
-      multiple: true
-      takes_value: true
-
-  - __ignored_lib_path:
-      short: L
-      takes_value: true
-      hidden: true
-      multiple: true
-      number_of_values: 1
-
-subcommands:
-  - print:
-      about: Prints NVPTX target definition JSON into stdout
-      args:
-        - TARGET:
-            help: Specifies the target name
-            required: true
-            possible_values: ["nvptx64-nvidia-cuda", "nvptx-nvidia-cuda"]
