+++ src/bin/legacy-rustc-dylib-wrapper/main.rs
