+++ targets/nvptx64-nvidia-cuda.json
@@ -1,8 +1,8 @@
 {
     "arch": "nvptx64",
     "cpu": "sm_20",
-    "data-layout": "e-p:64:64:64-i1:8:8-i8:8:8-i16:16:16-i32:32:32-i64:64:64-f32:32:32-f64:64:64-v16:16:16-v32:32:32-v64:64:64-v128:128:128-n16:32:64",
-    "linker": "ptx-linker",
+    "data-layout": "e-i64:64-i128:128-v16:16-v32:32-n16:32:64",
+    "linker": "legacy-ptx-linker",
     "linker-flavor": "ld",
     "linker-is-gnu": true,
     "dll-prefix": "",
